httpmetrics: provide standard middleware entrypoint
The httpmetrics package was confusing and awkward to use:

  1. It's only exported method was `NewServer`, which actually returned
     an `http.Handler`.
  2. It did not conform to standard middleware function signature, so it
     required every user to wrap in another function to use it.

This deprecates `NewServer` in favor of `New(metrics.Provider)` which
returns a normal middleware constructor.

It also moves the example into a `httpmetrics_test` package so that the
example code uses package names and is easier to drop into real code.
