+++ hmiddleware/httpmetrics/httpmetrics.go
@@ -14,46 +14,57 @@ import (
 	"github.com/heroku/x/go-kit/metricsregistry"
 )
 
+// New returns an HTTP middleware which captures request metrics and reports
+// them to the given provider.
+func New(p metrics.Provider) func(http.Handler) http.Handler {
+	reg := metricsregistry.New(p)
+	return func(next http.Handler) http.Handler {
+		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+			reg.GetOrRegisterCounter("http.server.all.requests").Add(1)
+
+			ww := middleware.NewWrapResponseWriter(w, r.ProtoMajor)
+
+			start := time.Now()
+			next.ServeHTTP(ww, r)
+			dur := time.Since(start)
+
+			st := ww.Status()
+			if st == 0 {
+				// Assume no Write or WriteHeader means OK.
+				st = http.StatusOK
+			}
+			sts := strconv.Itoa(st)
+
+			reg.GetOrRegisterHistogram("http.server.all.request-duration.ms", 50).Observe(ms(dur))
+			reg.GetOrRegisterCounter("http.server.all.response-statuses." + sts).Add(1)
+
+			ctx := r.Context()
+			if ctx.Value(chi.RouteCtxKey) == nil {
+				return
+			}
+			rtCtx := chi.RouteContext(ctx)
+
+			if len(rtCtx.RoutePatterns) == 0 {
+				// Did not match a route, give up.
+				return
+			}
+
+			// GET /apps/:foo/bars/:baz_id -> get.apps.foo.bars.baz-id
+			met := strings.ToLower(r.Method) + "." + nameRoutePatterns(rtCtx.RoutePatterns)
+			reg.GetOrRegisterCounter("http.server." + met + ".requests").Add(1)
+			reg.GetOrRegisterHistogram("http.server."+met+".request-duration.ms", 50).Observe(ms(dur))
+			reg.GetOrRegisterCounter("http.server." + met + ".response-statuses." + sts).Add(1)
+		})
+	}
+}
+
 // NewServer returns an http.Handler which calls next for each
 // request and reports metrics to the given provider.
+//
+// Deprecated: NewServer is awkward to use since it doesn't follow the normal
+// pattern for middleware. Use New instead.
 func NewServer(p metrics.Provider, next http.Handler) http.Handler {
-	reg := metricsregistry.New(p)
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		reg.GetOrRegisterCounter("http.server.all.requests").Add(1)
-
-		ww := middleware.NewWrapResponseWriter(w, r.ProtoMajor)
-
-		start := time.Now()
-		next.ServeHTTP(ww, r)
-		dur := time.Since(start)
-
-		st := ww.Status()
-		if st == 0 {
-			// Assume no Write or WriteHeader means OK.
-			st = http.StatusOK
-		}
-		sts := strconv.Itoa(st)
-
-		reg.GetOrRegisterHistogram("http.server.all.request-duration.ms", 50).Observe(ms(dur))
-		reg.GetOrRegisterCounter("http.server.all.response-statuses." + sts).Add(1)
-
-		ctx := r.Context()
-		if ctx.Value(chi.RouteCtxKey) == nil {
-			return
-		}
-		rtCtx := chi.RouteContext(ctx)
-
-		if len(rtCtx.RoutePatterns) == 0 {
-			// Did not match a route, give up.
-			return
-		}
-
-		// GET /apps/:foo/bars/:baz_id -> get.apps.foo.bars.baz-id
-		met := strings.ToLower(r.Method) + "." + nameRoutePatterns(rtCtx.RoutePatterns)
-		reg.GetOrRegisterCounter("http.server." + met + ".requests").Add(1)
-		reg.GetOrRegisterHistogram("http.server."+met+".request-duration.ms", 50).Observe(ms(dur))
-		reg.GetOrRegisterCounter("http.server." + met + ".response-statuses." + sts).Add(1)
-	})
+	return New(p)(next)
 }
 
 func ms(d time.Duration) float64 {
